Added to accounts summary scene and added new FXML file and controller
Created 2 BankAccount objects (A default chequing and savings account)
which both have a default amount of $0.00 indicated by labels to the
right of the names of the accounts and have buttons that can be clicked
on to view account details for a specific account. I also created and
linked a new FXML file and controller class for the purpose of showing a
separate scene for the account details of a particular account when the
button beside the account is clicked on in the accounts summary scene.